feat: begin nuxt 3 migration
