+++ .eslintrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ['@nuxtjs/eslint-config-typescript'],
+  rules: {
+    'vue/multi-word-component-names': 'off',
+    'import/named': 'off'
+  }
+}
