+++ lint-staged.config.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  '*.{js,ts,vue}': ['yarn lint']
+}
