+++ package.json
@@ -8,32 +8,60 @@
     "Davor Teskera",
     "Jonas Thelemann"
   ],
-  "keywords": [
-    "nuxt cookie control",
-    "nuxt gdpr",
-    "nuxt cookie consent",
-    "nuxt gdpr consent"
-  ],
-  "main": "lib/module.js",
-  "repository": "https://gitlab.com/dargmuesli/nuxt-cookie-control",
+  "exports": {
+    ".": {
+      "import": "./dist/module.mjs",
+      "require": "./dist/module.cjs"
+    }
+  },
+  "main": "./dist/module.cjs",
+  "types": "./dist/module.d.ts",
+  "repository": "https://github.com/dargmuesli/nuxt-cookie-control",
   "publishConfig": {
     "access": "public"
   },
-  "eslintIgnore": [
-    "*"
-  ],
   "files": [
-    "lib",
+    "dist",
     "components",
     "locale"
   ],
+  "scripts": {
+    "prepack": "nuxt-module-build",
+    "dev": "nuxi dev playground",
+    "dev:build": "nuxi build playground",
+    "dev:prepare": "nuxt-module-build --stub && nuxi prepare playground",
+    "lint": "eslint --ext .js,.ts,.vue",
+    "lint:all": "yarn lint .",
+    "release": "yarn test && release-it",
+    "test": "yarn vitest run"
+  },
   "dependencies": {
+    "@nuxt/kit": "3.0.0-rc.8",
     "css-vars-ponyfill": "2.4.8",
     "sass": "1.54.5",
     "sass-loader": "13.0.2",
     "string-replace-loader": "3.1.0"
   },
+  "devDependencies": {
+    "@nuxt/module-builder": "latest",
+    "@nuxt/test-utils": "npm:@nuxt/test-utils-edge@latest",
+    "@nuxtjs/eslint-config-typescript": "latest",
+    "@release-it/conventional-changelog": "5.1.0",
+    "c8": "7.12.0",
+    "consola": "2.15.3",
+    "eslint": "8.22.0",
+    "eslint-config-prettier": "8.5.0",
+    "nuxt": "2.15.8",
+    "release-it": "15.4.0",
+    "typescript": "4.5.5",
+    "vue": "3.2.37",
+    "webpack": "5.74.0"
+  },
   "peerDependencies": {
+    "vite": "^3",
     "webpack": "^5"
+  },
+  "resolutions": {
+    "@dargmuesli/nuxt-cookie-control": "link:./"
   }
 }
