+++ playground/nuxt.config.js
@@ -1,0 +1,5 @@
+import { defineNuxtConfig } from 'nuxt'
+
+export default defineNuxtConfig({
+  modules: ['@dargmuesli/nuxt-cookie-control']
+})
